Bug 24632: Adapt firefox and tor-browser to new macOS toolchain
After building Firefox we now get a 'Tor Browser.app' instead of a
'TorBrowser.app'. This patch makes sure the additional whitespace in the
app name is correctly handled by the build script and the one that deals
with packaging the final bundle.

We need to ship a fix for the Firefox packaging step as well as
|./mach package| wants to build the final .dmg in that step, too, which
breaks: there is no .dmg creation tool available. Setting
`INNER_MAKE_PACKAGE` to `true` does not seem to work anymore. That part
of this patch is currently only a workaround to get the nightly builds
going. We should come up with a better solution that allows us to skip
that part of the Firefox packaging step on all supported platforms.