Bug 26462: Unbreak Firefox 60 ESR 32bit Linux compilation
For some reason the switch from 60.0.1esr to 60.1.0esr is causing
to break our builds because the 32bit linker is not capable of linking
gkrust nor libxul due to address space limitations anymore.

We keep the configuration for 32bit Linux bundles as we shipped them in
our nightlis over the past couple of weeks but are avoiding building
browser debug information for now.