Bug 29319: Remove FTE support for Windows
We want to use our new mingw-w64/clang toolchain for our Windows
cross-compilation, which requires a newer Wine among other updated
components. We need to switch to Debian Stretch for that which is
necessary anyway sooner or later given that Jessie is not properly
maintained anymore.

Updating to Stretch breaks FTE and fixing the breakage seems
non-trivial. We therefore remove support for Windows where it existed
(there was no 64bit support implemented yet, see: #24195) as it is
planned to replace FTE with a new pluggable transport in the near future
anyway.