Bug 32976: Build and bundle geckodriver
Geckodriver is not build by default in release builds. However, it's useful
if we build it reproducibly so we can avoid trusting some binary downloaded
from the Internet for our QA and Selenium related work.

Only the 64bit Linux version is built and included as I doubt we'll need
geckodriver on anything else for our purposes.