Bug 32389: Sandbox Graphite using RLBox
Enable Firefox to use the WASM sandbox for Graphite on Linux x86_64.

We put all the patches rebased in #32380 into a separate directory in
our tor-browser-build repository (instead of having them on a, say,
tor-browser alpha branch, which feels more natural) to be able to apply
the partly invasive changes as fine-grained as possible. They should
only affect the targets we enable RLBox for, which is currently only
Linux x86_64 in the nightly channel.