Bug 32788: Specify version for test builds
With our switch to including a date into our nightly build versioning
(#25009) testing patches with test builds got harder. The firefox part
has to be rebuilt every day now as it includes the version number, even
if no code changes occurred. That slows down testing of patches
significantly.

We hardcode "testing" as the test build version number to fix that and
use "20010101010101" for the build date.