Bug 33626: Add GeckoView project
We remove the mobile parts from the firefox project and include them in
a new one, geckoview. While this increases the boilerplate parts a bit
it should make it easier to move mobile to the regular release train and
is conceptually clearer as well (no, we don't build firefox on mobile
but a library to be consumed by other projects).