Bug 33626: Add GeckoView project
We include a new project, geckoview. While this increases the
boilerplate a bit it should make it easier to move mobile to the regular
release series while desktop stays on ESR, and it is conceptually
clearer as well (no, we don't build firefox on mobile but a library to
be consumed by other projects).